Lots of fixes
Added the accessibility for the show/hide comments button, allowing it to be used with the tab and return keys instead of only the mouse. I also fixed up several additional HTML structure things I previously missed, such as adding <label> to a couple of things and changing the <td> elements into <th> ones.